+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 0.2.2
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Add Response::redirect() method to simplify redirects
+* Add Request::hasParam() method to allow query whether a param exists in the request or not
+
 # Version 0.2.1
 
 ## Bugfixes